move package correlator to common location
